new classes MonkeyToolbar and MonkeyStatusBar
